core: state: initialize: Find wallet commitment in Merkle history at startup
